pair mapping API, range move validation logic to it
